Huge Frontend Push (Read Description)
This push does a LOT.

First off, the sessionID API endpoints should be completed. On the initial render, getSessionIDs is called from the backend and renders in the SessionIDs component.

Secondly, when a sessionID is clicked on, the getTableFromSessionID is called by the frontend function fetchTableFromSessionID in TableDisplay.jsx, which gathers all the completed and uncompleted items connected to a sessionID in MongoDB.

Finally, I had to do probably the most debugging I've ever had to because I basically needed to render and use the calculateVendibility button regardless of how the current table data was retrieved (i.e. if the table was an uploaded excel or a table of data from MongoDB)